+++ src/code/const_init.rs
@@ -145,7 +145,7 @@ mod impl_devela_base_core {
         // data
         // media
         // num
-        f32bits, f64bits,
+        f32bits, f32bits_niche, f64bits, f64bits_niche,
         Cast, Cmp, Cycle, CycleCount, Interval, Sign,
         // text
         CharAscii, char7, char8, char16, charu, charu_niche,
@@ -158,7 +158,7 @@ mod impl_devela_base_core {
     pub use crate::{GraphemeNonul, GraphemeU8};
 
     // num
-    _impl_init![%Sealed%: f32bits, f64bits];
+    _impl_init![%Sealed%: f32bits, f32bits_niche, f64bits, f64bits_niche];
 
     _impl_init![%Sealed%: Sign];
     _impl_init![%Sealed%: <T: ConstInitCore> Cast<T>, Cmp<T>, Cycle<T>];